arm64: tegra: wire up PWM fan on Jetson TX1 DevKit base board
With enabling Bluetooth and WiFi the P2180 module starts to warm up, so
to keep it cool, enable PWM fan connected to J15 header on P2597 base
board. No colling map yet specified, as the polarity seems to be
inverted(?) and 255 value in sysfs is no rotation and 0 is the highest
rotation speed. Needs to be investigated.

Signed-off-by: Tomasz Maciej Nowak <tmn505@gmail.com>